pyproject.toml
[tool.poetry]
name = "bitcoin-trend-prediction"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
yfinance = "^0.2.43"
pandas = "^2.2.2"
scikit-learn = "^1.5.1"
mwclient = "^0.11.0"
transformers = "^4.44.2"
torch = "^2.4.1"
huggingface-hub = "^0.24.6"
filelock = "^3.15.4"
logging = "^0.4.9.6"
matplotlib = "^3.9.2"
boto3 = "^1.35.12"
numpy = "^2.1.1"
litserve = "^0.2.2"
python-dotenv = "^1.0.1"
xgboost = "^2.1.1"
joblib = "^1.4.2"
streamlit = "^1.38.0"
plotly = "^5.24.1"
azure-storage-blob = "^12.23.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
